# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for the encrypted-db project.",
        "title": "Encrypted-DB API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/currencies": {
            "post": {
                "description": "Create a new currency in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new currency",
                "parameters": [
                    {
                        "description": "Currency to create",
                        "name": "currency",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Currency"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "URL Scheme Error: Ensure URL scheme is 'http' or 'https'",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "201": {
                        "description": "Currency created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or duplicate symbol/contract",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create currency due to an unexpected error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/currencies/{hk}": {
            "put": {
                "description": "Update the information or status of an existing currency using its HK (UUID) as identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update a currency",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency HK",
                        "name": "hk",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Currency data to update",
                        "name": "currency",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Currency"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or duplicate symbol/contract",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Currency not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update currency due to an unexpected error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Logically delete a currency by setting deleted_at and changing its status to 'deleted'.",
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a currency",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency HK",
                        "name": "hk",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Currency not found or already deleted",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete currency due to an unexpected error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/public/currencies": {
            "get": {
                "description": "Retrieve all active currencies from Redis cache",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get all active currencies",
                "responses": {
                    "200": {
                        "description": "List of active currencies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Currency"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve currencies from cache",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/public/currencies/{hk}": {
            "get": {
                "description": "Retrieve a single currency from Redis cache using its HK",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get a single currency by HK",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency HK",
                        "name": "hk",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency data",
                        "schema": {
                            "$ref": "#/definitions/models.Currency"
                        }
                    },
                    "404": {
                        "description": "Currency not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve currency from cache",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "info": {
                    "description": "Info can hold any type of data (e.g., list or object)"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Currency": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "hk": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "$ref": "#/definitions/models.CurrencyInfo"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CurrencyInfo": {
            "type": "object",
            "properties": {
                "contract": {
                    "type": "string"
                },
                "en_name": {
                    "type": "string"
                },
                "fa_name": {
                    "type": "string"
                },
                "if_fiat": {
                    "type": "boolean"
                },
                "is_token": {
                    "type": "boolean"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}